- Fixed a few places where we had int16_t vs uint16t - Fixed a few places with bits that would not work with 8 bit compilers - Fixed a few places to typecast between an int and uint16_t in loops for the 8 bit compilers - The code compiles for a PIC18F but the XC8 compiler is very sophisticated and analyzes the entire project.  With the extensive use of function pointers in the library it has a hard time with detecting recursion and since the PIC parts have limits to the stack it throws a warning about unknown recursion and warnings about stack overflow.

+++ src/openlcb/openlcb_utilities.c
@@ -352,7 +352,7 @@ bool OpenLcbUtilities_is_producer_event_assigned_to_node(openlcb_node_t* openlcb
 
         if (openlcb_node->producers.list[i].event == event_id) {
 
-            (*event_index) = i;
+            (*event_index) = (uint16_t) i;
 
             return true;
 
@@ -370,7 +370,7 @@ bool OpenLcbUtilities_is_consumer_event_assigned_to_node(openlcb_node_t* openlcb
 
         if (openlcb_node->consumers.list[i].event == event_id) {
 
-            (*event_index) = i;
+            (*event_index) = (uint16_t) i;
 
             return true;
